dts/arm: STM32: Removing i2s1 node from stm32f4 series.
This commit removes the i2s1 node from stm32f4 series as i2s1
does not exist in stm32f4 series. The sequence starts from i2s2 onwards.
Also this i2s1 node was having the same addresses and clock as that
of spi1 which is not correct as the reference manual.

Signed-off-by: Krishna Mohan Dani <krishnamohan.d@hcl.com>